python-worker/requirements.txt
# DBOS Python SDK
dbos>=0.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Object Detection
ultralytics>=8.0.0  # YOLOv8
opencv-python>=4.8.0
Pillow>=10.0.0

# OCR (for future implementation)
# paddleocr>=2.7.0
# paddlepaddle>=2.5.0

# HTTP client
requests>=2.31.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0  # PostgreSQL driver for intent polling

# Simple-workflow library
# Note: Loaded from ../simple-workflow/python via sys.path in main.py
# Requires: psycopg2-binary (listed above)
